refactor(view): 입력 및 출력 관련 함수 분리 - 입력 및 출력 함수를 RaceGame 클래스에서 InputView, OutputView 클래스로 분리함. - RaceGame에서 InputView, OutputView 객체를 생성하도록 변경함. - 게임 시작 함수였던 play() 함수 이름을 playGame()으로 변경 - 결과 출력 함수를 endGame() 메서드 내에 감싸 분리하였다.
