+++ docs/README.md
@@ -1,26 +1,19 @@
 ## 요구 사항 명세서(SRS)
 ### 1. 어떤 기능들이 있는가?
-- [ ] 자동차(Car)의 이름을 입력받는다.(* 이름은 5자 이하)
-- [ ] 입력한 자동차들을 쉼표(,)를 기준으로 경주 게임에 참가시킨다.
-- [ ] 참가한 자동차가 최소 2대 이상 있어야 한다.
-- [ ] 몇 번의 이동을 할 것인지 사용자로부터 입력받는다.
-- [ ] 각 자동차는 매 경기 차수마다 무작위 숫자를 발생시킨다.(* 범위 0~9)
-- [ ] 발생시킨 숫자가 4이상일 경우 해당 자동차는 1칸씩 이동한다. 그렇지 않을 경우 멈춰있는다.
-- [ ] 모든 차수의 경기가 끝난 후 경기 결과를 확인 후 우승자의 이름을 출력한다.(* 우승자 여러명 가능)
-- [ ] 사용자가 잘못된 입력을 할 경우 `IllegalArgumentException`을 발생시킨 후 어플리케이션은 종료된다.
+- [x] 자동차(Car)의 이름을 입력받는다.(* 이름은 5자 이하)
+- [x] 입력한 자동차들을 쉼표(,)를 기준으로 경주 게임에 참가시킨다.
+- [x] 참가한 자동차가 최소 2대 이상 있어야 한다.
+- [x] 몇 번의 이동을 할 것인지 사용자로부터 입력받는다.
+- [x] 각 자동차는 매 경기 차수마다 무작위 숫자를 발생시킨다.(* 범위 0~9)
+- [x] 발생시킨 숫자가 4이상일 경우 해당 자동차는 1칸씩 이동한다. 그렇지 않을 경우 멈춰있는다.
+- [x] 모든 차수의 경기가 끝난 후 경기 결과를 확인 후 우승자의 이름을 출력한다.(* 우승자 여러명 가능)
+- [x] 사용자가 잘못된 입력을 할 경우 `IllegalArgumentException`을 발생시킨 후 어플리케이션은 종료된다.
 ---
 ### 2. 어떤 객체들로 이루어져있는가?
 2.1. `[RaceGame]`: 게임 클래스
    - [x] 게임을 실행한다.
-   - [x] 사용자의 입력에 대한 유효성 검사: 경주할 차가 최소 2대 이상인지 확인
-   - [x] 사용자의 입력에 대한 유효성 검사: 경주할 차의 이름의 길이가 최소 1자이상 최대 5자 이하인지 확인
-   - [x] 입력받은 참가자들을 기록한다.
-   - [x] 이동횟수를 입력받는다.
-   - [x] 사용자의 입력에 대한 유효성 검사: 이동 횟수가 숫자인지 확인한다.
-   - [x] 사용자의 입력에 대한 유효성 검사: 이동 횟수가 1이상인지 확인한다.
    - [x] 입력받은 횟수만큼 매 라운드를 실행한다.
-   - [x] 모든 라운드가 종료된 이후 우승자를 출력한다.
-   - [ ] 차 이름의 앞과 뒤 공백이 있을 경우 제거하고 유효성 검사 함수를 호출한다.
+   - [x] 모든 라운드가 종료된 이후 우승자를 출력함수를 호출한다.
 
 2.2. `[Car]` : 자동차 클래스
    - [x] 자동차 이름을 가진다.
@@ -35,6 +28,17 @@
    - [x] 참가한 자동차가 최소 2대 이상인지 확인한다.
    - [x] 이동 횟수는 문자가 아닌 숫자이어야 한다.
    - [x] 이동 횟수는 최소 1이상의 값이어야 한다.
+
+2.4 `[InputView]` : 입력 클래스
+- [x] 사용자의 입력에 대한 유효성 검사: 경주할 차가 최소 2대 이상인지 확인
+- [x] 사용자의 입력에 대한 유효성 검사: 경주할 차의 이름의 길이가 최소 1자이상 최대 5자 이하인지 확인
+- [x] 입력받은 참가자들을 기록한다.
+- [x] 이동횟수를 입력받는다.
+- [x] 사용자의 입력에 대한 유효성 검사: 이동 횟수가 숫자인지 확인한다.
+- [x] 사용자의 입력에 대한 유효성 검사: 이동 횟수가 1이상인지 확인한다.
+
+2.5 `[OutputView]` : 출력 클래스
+- [x] 우승자를 출력한다.
 ---
 ### 3. 어떻게 상호작용하는가?
 3.1. `RaceGame` 클래스
