+++ src/main/java/racingcar/utils/InputValidator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.utils;
 
 import java.util.List;
